requirements.txt
numpy 
osmnx
networkx
Pillow
imageio
